added simple ranking

+++ bots/cowculator.h
@@ -19,7 +19,7 @@ class Cowculator : public Bot
 private:
   GameInfo gameInfo;
   std::pair<int, int> move(const GameState& state, int timerOffset);
-  int calculateRankOfState(const GameState& state);
+  double calculateRankOfState(const GameState& state);
   GameState processGameState(const GameState& state);
   GameState approximateNextState(const GameState& state, int timeDelta);
   MovingObject movePaddle(const MovingObject& obj, int timeDelta);
